feat(ios): add TTS test button and fix headset to pause command
- Add 'Test Voice' button in OpenClaw settings that speaks a sample
  phrase using current voice/model/speed and shows errors inline
- Switch headset acknowledge from togglePlayPause to pauseCommand only
  (more reliable on AirPods; single-tap was not triggering)
- Update setting label to 'Headset Pause Acknowledge'
- Extract OpenClawSettingsView into its own file to fix file_length lint
